Move Ignore and Throw from test to prod code
