+++ elastic-job-lite-core/src/main/java/org/apache/shardingsphere/elasticjob/lite/executor/handler/error/JobExceptionHandlerFactory.java
@@ -34,7 +34,7 @@ public final class JobExceptionHandlerFactory {
     
     private static final Map<String, JobExceptionHandler> HANDLERS = new LinkedHashMap<>();
     
-    private static final String DEFAULT_HANDLER = "LOG_ERROR";
+    private static final String DEFAULT_HANDLER = "LOG";
     
     static {
         for (JobExceptionHandler each : ServiceLoader.load(JobExceptionHandler.class)) {
