+++ elastic-job-lite-core/src/main/java/org/apache/shardingsphere/elasticjob/lite/executor/handler/error/impl/ThrowJobExceptionHandler.java
@@ -7,7 +7,7 @@
  * the License.  You may obtain a copy of the License at
  *
  *     http://www.apache.org/licenses/LICENSE-2.0
- *
+ *  
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -15,18 +15,23 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.elasticjob.lite.fixture.handler;
+package org.apache.shardingsphere.elasticjob.lite.executor.handler.error.impl;
 
+import org.apache.shardingsphere.elasticjob.lite.exception.JobSystemException;
 import org.apache.shardingsphere.elasticjob.lite.executor.handler.error.JobExceptionHandler;
 
-public final class IgnoreJobExceptionHandler implements JobExceptionHandler {
+/**
+ * Job exception handler for throw exception.
+ */
+public final class ThrowJobExceptionHandler implements JobExceptionHandler {
     
     @Override
     public void handleException(final String jobName, final Throwable cause) {
+        throw new JobSystemException(cause);
     }
     
     @Override
     public String getType() {
-        return "IGNORE";
+        return "THROW";
     }
 }
