+++ elastic-job-lite-core/src/main/resources/META-INF/services/org.apache.shardingsphere.elasticjob.lite.executor.handler.error.JobExceptionHandler
@@ -15,4 +15,6 @@
 # limitations under the License.
 #
 
-org.apache.shardingsphere.elasticjob.lite.executor.handler.error.impl.LogErrorJobExceptionHandler
+org.apache.shardingsphere.elasticjob.lite.executor.handler.error.impl.LogJobExceptionHandler
+org.apache.shardingsphere.elasticjob.lite.executor.handler.error.impl.IgnoreJobExceptionHandler
+org.apache.shardingsphere.elasticjob.lite.executor.handler.error.impl.ThrowJobExceptionHandler
