+++ elastic-job-lite-spring/src/test/resources/META-INF/job/withJobHandler.xml
@@ -33,5 +33,5 @@
                 registry-center-ref="regCenter" cron="${simpleJob.cron}" sharding-total-count="${simpleJob.shardingTotalCount}" sharding-item-parameters="${simpleJob.shardingItemParameters}" 
                 disabled="${simpleJob.disabled}" overwrite="${simpleJob.overwrite}" job-executor-service-handler="SIMPLE" />
     <job:dataflow id="dataflowElasticJob_namespace_job_handler" job-ref="dataflowJob" registry-center-ref="regCenter" cron="0/1 * * * * ?" sharding-total-count="3" sharding-item-parameters="0=A,1=B,2=C" 
-                  description="中文描述" overwrite="true" job-exception-handler="SIMPLE" />
+                  description="中文描述" overwrite="true" job-exception-handler="THROW" />
 </beans>
